refactor: organize the components in a tree
